basic idea of moving api calls to a util
